+++ ZKSyncSDK/Classes/Model/Token/Token.swift
@@ -13,7 +13,7 @@ public struct Token: TokenId, Decodable {
     
     private static let DefaultAddress = "0x0000000000000000000000000000000000000000"
     
-    let id: UInt16
+    let id: UInt32
     public let address: String
     let symbol: String
     let decimals: Int
