+++ ZKSyncSDK/Classes/Model/Transactions/MintNFT.swift
@@ -17,14 +17,14 @@ public class MintNFT: ZkSyncTransaction {
     let contentHash: String
     let recipient: String
     let fee: String
-    let feeToken: UInt16
+    let feeToken: UInt32
     let nonce: UInt32
     
     var signature: Signature?
 
     var feeInteger: BigUInt { BigUInt(fee)! }
 
-    public init(creatorId: UInt32, creatorAddress: String, contentHash: String, recipient: String, fee: String, feeToken: UInt16, nonce: UInt32, signature: Signature? = nil) {
+    public init(creatorId: UInt32, creatorAddress: String, contentHash: String, recipient: String, fee: String, feeToken: UInt32, nonce: UInt32, signature: Signature? = nil) {
         self.creatorId = creatorId
         self.creatorAddress = creatorAddress
         self.contentHash = contentHash
