+++ ZKSyncSDK/Classes/Model/Transactions/Withdraw.swift
@@ -15,7 +15,7 @@ public class Withdraw: ZkSyncTransaction {
     let accountId: UInt32
     let from: String
     let to: String
-    let token: UInt16
+    let token: UInt32
     let amount: BigUInt
     let fee: String
     let nonce: UInt32
@@ -25,7 +25,7 @@ public class Withdraw: ZkSyncTransaction {
 
     var feeInteger: BigUInt { BigUInt(fee)! }
 
-    public init(accountId: UInt32, from: String, to: String, token: UInt16, amount: BigUInt, fee: String, nonce: UInt32, timeRange: TimeRange) {
+    public init(accountId: UInt32, from: String, to: String, token: UInt32, amount: BigUInt, fee: String, nonce: UInt32, timeRange: TimeRange) {
         self.accountId = accountId
         self.from = from
         self.to = to
