+++ ZKSyncSDK/Classes/Signer/Ethereum/EthSigner.swift
@@ -28,7 +28,7 @@ public protocol EthSigner {
     
     func signMintNFT(contentHash: String, recepient: String, nonce: UInt32, token: Token, fee: BigUInt) throws -> EthSignature
 
-    func signWithdrawNFT(to: String, tokenId: UInt16, nonce: UInt32, token: Token, fee: BigUInt) throws -> EthSignature
+    func signWithdrawNFT(to: String, tokenId: UInt32, nonce: UInt32, token: Token, fee: BigUInt) throws -> EthSignature
     
     func sign(message: Data) throws -> EthSignature
